Move Create Table to own file
